refactor: 함수의 역할 분리
- 메서드 분리
- commerse -> commerce 로 오타 수정
