+++ README.md
@@ -1,1 +1,1 @@
-# E-Commerse
+# E-Commerce
