+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'commerse'
+rootProject.name = 'commerce'
