+++ src/main/java/com/maljaaa/commerce/CommerceApplication.java
@@ -1,0 +1,13 @@
+package com.maljaaa.commerce;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CommerceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CommerceApplication.class, args);
+	}
+
+}
