+++ src/main/java/com/maljaaa/commerce/controller/AuthController.java
@@ -1,0 +1,36 @@
+package com.maljaaa.commerce.controller;
+
+import com.maljaaa.commerce.dto.MemberRequestDto;
+import com.maljaaa.commerce.dto.MemberResponseDto;
+import com.maljaaa.commerce.dto.TokenDto;
+import com.maljaaa.commerce.dto.TokenRequestDto;
+import com.maljaaa.commerce.service.AuthService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/auth")
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<MemberResponseDto> signup(@RequestBody MemberRequestDto memberRequestDto) {
+        return ResponseEntity.ok(authService.signup(memberRequestDto));
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<TokenDto> login(@RequestBody MemberRequestDto memberRequestDto) {
+        return ResponseEntity.ok(authService.login(memberRequestDto));
+    }
+
+    @PostMapping("/reissue")
+    public ResponseEntity<TokenDto> reissue(@RequestBody TokenRequestDto tokenRequestDto) {
+        return ResponseEntity.ok(authService.reissue(tokenRequestDto));
+    }
+}
