+++ src/main/java/com/maljaaa/commerce/controller/ViewController.java
@@ -1,0 +1,15 @@
+package com.maljaaa.commerce.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+@Slf4j
+public class ViewController {
+    @GetMapping("/")
+    public String index() {
+        log.info("View Controller");
+        return "index.html";
+    }
+}
