+++ src/main/java/com/maljaaa/commerce/dto/TokenDto.java
@@ -1,0 +1,16 @@
+package com.maljaaa.commerce.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+@Builder
+public class TokenDto {
+
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+    private Long accessTokenExpiresIn;
+}
