+++ src/main/java/com/maljaaa/commerce/dto/TokenRequestDto.java
@@ -1,0 +1,12 @@
+package com.maljaaa.commerce.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class TokenRequestDto {
+
+    private String accessToken;
+    private String refreshToken;
+}
