+++ src/main/java/com/maljaaa/commerce/model/Authority.java
@@ -1,0 +1,5 @@
+package com.maljaaa.commerce.model;
+
+public enum Authority {
+    ROLE_USER, ROLE_ADMIN
+}
