+++ src/main/java/com/maljaaa/commerce/model/jwt/RefreshToken.java
@@ -1,0 +1,33 @@
+package com.maljaaa.commerce.model.jwt;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@Getter
+public class RefreshToken {
+
+    @Id
+    @Column(name = "rf_key")
+    private String key;
+
+    @Column(name = "rt_value")
+    private String value;
+
+    @Builder
+    public RefreshToken(String key, String value) {
+        this.key = key;
+        this.value = value;
+    }
+
+    public RefreshToken updateValue(String token) {
+        this.value = token;
+        return this;
+    }
+
+}
