+++ src/main/java/com/maljaaa/commerce/repository/jwt/RefreshTokenRepository.java
@@ -1,0 +1,10 @@
+package com.maljaaa.commerce.repository.jwt;
+
+import com.maljaaa.commerce.model.jwt.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByKey(String key);
+}
