--- src/main/java/com/maljaaa/commerse/CommerseApplication.java
@@ -1,13 +1,0 @@
-package com.maljaaa.commerse;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CommerseApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CommerseApplication.class, args);
-	}
-
-}
