--- src/main/java/com/maljaaa/commerse/controller/ViewController.java
@@ -1,15 +1,0 @@
-package com.maljaaa.commerse.controller;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-@Slf4j
-public class ViewController {
-    @GetMapping("/")
-    public String index() {
-        log.info("View Controller");
-        return "index.html";
-    }
-}
