--- src/main/java/com/maljaaa/commerse/dto/LoginDto.java
@@ -1,21 +1,0 @@
-package com.maljaaa.commerse.dto;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-@Builder
-public class LoginDto {
-
-    @NotNull
-    @Size(min = 3, max = 50)
-    private String name;
-
-    @NotNull
-    @Size(min = 3, max = 100)
-    private String pwd;
-}
