--- src/main/java/com/maljaaa/commerse/dto/MemberResponseDto.java
@@ -1,16 +1,0 @@
-package com.maljaaa.commerse.dto;
-
-import com.maljaaa.commerse.model.Member;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class MemberResponseDto {
-
-    private String name;
-
-    public static MemberResponseDto of(Member member) {
-        return new MemberResponseDto(member.getEmail());
-    }
-}
