--- src/main/java/com/maljaaa/commerse/dto/TokenDto.java
@@ -1,16 +1,0 @@
-package com.maljaaa.commerse.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-@Builder
-public class TokenDto {
-
-    private String grantType;
-    private String accessToken;
-    private String refreshToken;
-    private Long accessTokenExpiresIn;
-}
