--- src/main/java/com/maljaaa/commerse/dto/TokenRequestDto.java
@@ -1,12 +1,0 @@
-package com.maljaaa.commerse.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class TokenRequestDto {
-
-    private String accessToken;
-    private String refreshToken;
-}
