--- src/main/java/com/maljaaa/commerse/dto/login/LoginUserRequest.java
@@ -1,7 +1,0 @@
-package com.maljaaa.commerse.dto.login;
-
-import lombok.Data;
-
-@Data
-public class LoginUserRequest {
-}
