--- src/main/java/com/maljaaa/commerse/dto/login/LoginUserResponse.java
@@ -1,7 +1,0 @@
-package com.maljaaa.commerse.dto.login;
-
-import lombok.Data;
-
-@Data
-public class LoginUserResponse {
-}
