--- src/main/java/com/maljaaa/commerse/dto/signup/CreateUserRequest.java
@@ -1,17 +1,0 @@
-package com.maljaaa.commerse.dto.signup;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Data;
-
-@Data
-public class CreateUserRequest {
-
-    @NotEmpty
-    private String name;
-
-    @NotEmpty
-    private String email;
-
-    @NotEmpty
-    private String pwd;
-}
