--- src/main/java/com/maljaaa/commerse/dto/signup/CreateUserResponse.java
@@ -1,17 +1,0 @@
-package com.maljaaa.commerse.dto.signup;
-
-import lombok.Data;
-
-import static com.maljaaa.commerse.utils.SuccessMessage.*;
-
-@Data
-public class CreateUserResponse {
-
-    private Long id;
-    private String message;
-
-    public CreateUserResponse(Long id, String name) {
-        this.id = id;
-        this.message = name + SIGNUP_OK.getMessage();
-    }
-}
