--- src/main/java/com/maljaaa/commerse/jwt/JwtAccessDeniedHandler.java
@@ -1,21 +1,0 @@
-package com.maljaaa.commerse.jwt;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-import static jakarta.servlet.http.HttpServletResponse.*;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        response.sendError(SC_FORBIDDEN);
-    }
-}
