--- src/main/java/com/maljaaa/commerse/jwt/JwtAuthenticationEntryPoint.java
@@ -1,20 +1,0 @@
-package com.maljaaa.commerse.jwt;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-import static jakarta.servlet.http.HttpServletResponse.*;
-
-@Component
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.sendError(SC_UNAUTHORIZED);
-    }
-}
