--- src/main/java/com/maljaaa/commerse/model/Authority.java
@@ -1,5 +1,0 @@
-package com.maljaaa.commerse.model;
-
-public enum Authority {
-    ROLE_USER, ROLE_ADMIN
-}
