--- src/main/java/com/maljaaa/commerse/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.maljaaa.commerse.repository;
-
-import com.maljaaa.commerse.model.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    List<Member> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
