--- src/main/java/com/maljaaa/commerse/repository/jwt/RefreshTokenRepository.java
@@ -1,10 +1,0 @@
-package com.maljaaa.commerse.repository.jwt;
-
-import com.maljaaa.commerse.model.jwt.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByKey(String key);
-}
