--- src/main/java/com/maljaaa/commerse/service/CustomUserDetailService.java
@@ -1,40 +1,0 @@
-package com.maljaaa.commerse.service;
-
-import com.maljaaa.commerse.model.Member;
-import com.maljaaa.commerse.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collections;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailService implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return memberRepository.findByEmail(username)
-                .stream().map(this::createUserDetails)
-                .findAny().orElseThrow(() -> new UsernameNotFoundException(username + " -> DB에서 찾을 수 없습니다."));
-    }
-
-    private UserDetails createUserDetails(Member member) {
-        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(member.getAuthority().toString());
-
-        return new User(
-                String.valueOf(member.getEmail()),
-                member.getPwd(),
-                Collections.singleton(grantedAuthority)
-        );
-    }
-}
