+++ src/test/java/com/maljaaa/commerce/CommerceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.maljaaa.commerce;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommerceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
