--- src/test/java/com/maljaaa/commerse/CommerseApplicationTests.java
@@ -1,13 +1,0 @@
-package com.maljaaa.commerse;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommerseApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
